fix: move types.ts into utils
