+++ src/components/Footer/Footer.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import logoRSS from '../../assets/rss.svg';
-import { Authors, AuthorsLink } from '../../types';
+import { Authors, AuthorsLink } from '../../utils/types';
 
 import style from './Footer.module.scss';
